Be more carefull when cleaning up certificates
Avoid cleaning up the current certificate, even if it's expired. This
ensures there is always a certifcate for services to use (even if
invalid).

Also don't cleanup certificates if they have a NotBefore time in the
future. As this is likely caused by an issue with the local clock (e.g.
incorrect RTC setting at boot), rather then the certificate being
invalid.